Move Hive transactional tables tests to separate class
This also updates the exception message.
