+++ presto-product-tests/src/main/java/io/prestosql/tests/hive/TestInsertIntoHiveTable.java
@@ -35,7 +35,6 @@ import static io.prestosql.tempto.fulfillment.table.TableRequirements.immutableT
 import static io.prestosql.tempto.fulfillment.table.TableRequirements.mutableTable;
 import static io.prestosql.tempto.query.QueryExecutor.query;
 import static io.prestosql.tests.hive.AllSimpleTypesTableDefinitions.ALL_HIVE_SIMPLE_TYPES_TEXTFILE;
-import static io.prestosql.tests.utils.QueryExecutors.onHive;
 
 public class TestInsertIntoHiveTable
         extends ProductTest
@@ -146,18 +145,5 @@ public class TestInsertIntoHiveTable
         String tableNameInDatabase = mutableTablesState().get(PARTITIONED_TABLE_WITH_SERDE).getNameInDatabase();
         assertThat(query("INSERT INTO " + tableNameInDatabase + " SELECT 1, 'presto', '2018-01-01'")).containsExactly(row(1));
         assertThat(query("SELECT * FROM " + tableNameInDatabase)).containsExactly(row(1, "presto", "2018-01-01"));
-    }
-
-    @Test
-    public void testInsertIntoTransactionalTable()
-    {
-        onHive().executeQuery("" +
-                "CREATE TABLE test_insert_into_transactional_table(a bigint)" +
-                "CLUSTERED BY(a) INTO 4 BUCKETS STORED AS ORC TBLPROPERTIES ('transactional'='true', 'bucketing_version'='1')");
-
-        assertThat(() -> query("INSERT INTO test_insert_into_transactional_table (a) VALUES (42)"))
-                .failsWithMessage("Inserting into Hive transactional tables is not supported: default.test_insert_into_transactional_table");
-
-        onHive().executeQuery("DROP TABLE test_insert_into_transactional_table");
     }
 }
